# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexFlow CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signals
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
  /signals/{id}/dismiss:
    post:
      operationId: dismissSignal
      tags: [signals]
      summary: Dismiss a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signal dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"

  # ── Segments ──
  /segments:
    get:
      operationId: listSegments
      tags: [segments]
      summary: List segments
      responses:
        "200":
          description: Segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Segment"
    post:
      operationId: createSegment
      tags: [segments]
      summary: Create segment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSegmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
  /segments/{id}:
    get:
      operationId: getSegment
      tags: [segments]
      summary: Get segment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
    delete:
      operationId: deleteSegment
      tags: [segments]
      summary: Delete segment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Calls ──
...
components:
  schemas:
...
    Signal:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        contact_id:
          type: string
        contact_name:
          type: string
        company_id:
          type: string
        company_name:
          type: string
        type:
          type: string
          enum: [funding_round, exec_move, expansion, hiring, product_launch, news, social]
        title:
          type: string
        body:
          type: string
        score:
          type: number
        status:
          type: string
          enum: [new, viewed, actioned, dismissed]
        source_url:
          type: string
        created_at:
          type: string
      required:
        - id
        - type
        - title
        - score
        - status
        - created_at
    CreateSignalBody:
      type: object
      properties:
        contact_id:
          type: string
        company_id:
          type: string
        type:
          type: string
          enum: [funding_round, exec_move, expansion, hiring, product_launch, news, social]
        title:
          type: string
        body:
          type: string
        source_url:
          type: string
      required:
        - type
        - title
    Segment:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        description:
          type: string
        filter_query:
          type: string
        contact_count:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - created_at
    CreateSegmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        filter_query:
          type: string
      required:
        - name
...